requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.7.2
idna==2.10
multidict==4.7.6
psycopg2==2.8.5
typing-extensions==3.7.4.2
websockets==8.1
yarl==1.5.1
